YG-81 💡 floatingBar 파일 분리
