+++ app/(afterLogin)/detailPost/[postId]/page.tsx
@@ -14,16 +14,10 @@ import CreateComment from "./_components/comment/createComment"
 import LikeToComment from "./_components/comment/likeToComment"
 import CommentBox from "./_components/comment/commentBox"
 import { Comment } from "./_components/comment/type"
-import { useRouter } from "next/navigation"
-import { useDeletePost } from "@/hook/usePostMutation"
-import { usePostDataStore } from "@/libs/store"
 import { defaultIcons, handlePostIcons } from "@/constants/floatingBarIcons"
 
 const DetailPostPage = ({ params }: PostDetailProps) => {
     const { postId } = params
-    const router = useRouter()
-    const deletePostMutation = useDeletePost()
-    const { setPostId, setPostDetail } = usePostDataStore()
 
     const {
         data: post,
@@ -43,20 +37,6 @@ const DetailPostPage = ({ params }: PostDetailProps) => {
         queryFn: () => getComment({ postId: Number(postId) }),
     })
 
-    const handleDeletePost = async () => {
-        try {
-            await deletePostMutation.mutateAsync(Number(postId))
-        } catch {
-            // 삭제 실패 오버레이 적용
-        }
-    }
-
-    const handleUpdatePost = (post: Post) => {
-        setPostId(postId)
-        setPostDetail(post)
-        router.push(`/updatePost/freeForm/${postId}`)
-    }
-
     if (isLoading || isCommentLoading) return <div>Loading...</div>
     if (error) return <div>Error: {error.message}</div>
     if (commentError) return <div>Error: {commentError.message}</div>
@@ -66,8 +46,7 @@ const DetailPostPage = ({ params }: PostDetailProps) => {
         <div className="flex items-center justify-center flex-col">
             <div className="relative w-[1300px] flex flex-col items-center justify-center  pt-10 ">
                 <FloatingBar icons={defaultIcons} />
-                <FloatingBar icons={handlePostIcons} isMine={true} />
-
+                <FloatingBar icons={handlePostIcons} isMine={true} postId={postId} post={post} />
                 {post.content !== "" ? (
                     <FreeFormDetail
                         title={post.title}
