+++ hook/useHandleScroll.ts
@@ -1,0 +1,20 @@
+import { useState, useEffect } from "react"
+
+const useHandleScroll = () => {
+    const [scrollY, setScrollY] = useState(0)
+
+    const handleScroll = () => {
+        setScrollY(window.scrollY)
+    }
+
+    useEffect(() => {
+        window.addEventListener("scroll", handleScroll)
+        return () => {
+            window.removeEventListener("scroll", handleScroll)
+        }
+    }, [])
+
+    return scrollY
+}
+
+export default useHandleScroll
